Actually we do need to hash in cargo
