+++ crates/paralegal-flow/src/args.rs
@@ -16,17 +16,11 @@
 use anyhow::Error;
 use clap::ValueEnum;
 use rustc_data_structures::fx::{FxHashMap, FxHashSet};
-use rustc_data_structures::stable_hasher::{HashStable, StableHasher};
 use rustc_hir::def_id::{CrateNum, LOCAL_CRATE};
-use rustc_macros::HashStable;
-use rustc_middle::ich::StableHashingContext;
 use rustc_middle::ty::TyCtxt;
 use rustc_span::Symbol;
-use rustc_stable_hash::ExtendedHasher;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::ffi::{OsStr, OsString};
-use std::hash::{Hash, Hasher};
-use std::path::Path;
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::sync::OnceLock;
@@ -241,28 +235,6 @@ impl Args {
     pub fn setup_logging(&self) -> anyhow::Result<()> {
         setup_logging()
     }
-
-    pub fn hash_config(&self, hasher: &mut impl Hasher) {
-        if self.attach_to_debugger.is_some() {
-            // If we run the debugger try to make the hash fail so we actually run.
-            std::time::Instant::now().hash(hasher);
-        }
-        // TODO Add other relevant arguments
-        config_hash_for_file(self.build_config.0.as_ref(), hasher);
-        self.relaxed.hash(hasher);
-        self.target.hash(hasher);
-        self.result_path.hash(hasher);
-        config_hash_for_file(self.marker_control.external_annotations().as_ref(), hasher);
-    }
-}
-
-fn config_hash_for_file(path: Option<impl AsRef<Path>>, state: &mut impl Hasher) {
-    path.as_ref()
-        .map(|path| {
-            let path = path.as_ref();
-            (path, path.metadata().unwrap().modified().unwrap())
-        })
-        .hash(state);
 }
 
 pub struct AnalysisCtrl {
